kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/fast/events/touch/touch-child-inside-empty-element.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="./resources/touch-util.js"></script>
<style>

#test1 {
    float: left; /* makes the container height 0 */
    width: 100px;
    height: 100px;
    background-color: green;
}

</style>
</head>
<body>
<div style="position: absolute">
  <!-- test0 element will have the touch handler and height 0 -->
  <div id="test0">
    <!-- test1 element will be the target of the touch -->
    <div id="test1"></div>
  </div>
</div>

<p id="description"></p>
<div id="console"></div>

<script type="text/javascript">
var lastEvent = null;
var touchEventsReceived = 0;
var EXPECTED_TOUCH_EVENTS_TOTAL = 3;

function touchEventCallback(event) {
    if (window.eventSender) {
        lastEvent = event;
        verifyTouch(touchEventsReceived++);
    } else {
        debug(event.type);
    }

    if (window.layoutTestController && touchEventsReceived == EXPECTED_TOUCH_EVENTS_TOTAL) {
        // If we've got here, we can safely say we were successfully parsed :) We need to
        // call the isSucccessfullyParsed function to output the correct TEST COMPLETE
        // footer message.
        isSuccessfullyParsed();
        layoutTestController.notifyDone();
    }
}

var div = document.getElementById("test0");
div.addEventListener("touchstart", touchEventCallback, false);
div.addEventListener("touchmove", touchEventCallback, false);
div.addEventListener("touchend", touchEventCallback, false);

function verifyTouchEvent(type, totalTouchCount, changedTouchCount, targetTouchCount)
{
    shouldBeEqualToString("lastEvent.type", type);
    shouldBe("lastEvent.touches.length", totalTouchCount.toString());
    shouldBe("lastEvent.changedTouches.length", changedTouchCount.toString());
    shouldBe("lastEvent.targetTouches.length", targetTouchCount.toString());
    shouldBe("lastEvent.pageX", "0");
    shouldBe("lastEvent.pageY", "0");
}

function verifyTouchPoint(list, point, x, y, id)
{
    shouldBe("lastEvent." + list + "[" + point + "].pageX", x.toString());
    shouldBe("lastEvent." + list + "[" + point + "].pageY", y.toString());
    shouldBe("lastEvent." + list + "[" + point + "].clientX", x.toString());
    shouldBe("lastEvent." + list + "[" + point + "].clientY", y.toString());
    shouldBe("lastEvent." + list + "[" + point + "].identifier", id.toString());
}

function verifyTouch(which) {
    switch (which) {
    case 0:
        verifyTouchEvent("touchstart", 1, 1, 1);
        shouldBeEqualToString("lastEvent.touches[0].target.id", "test1");
        verifyTouchPoint("touches", 0, 100, 50, 0);
        verifyTouchPoint("changedTouches", 0, 100, 50, 0);
        verifyTouchPoint("targetTouches", 0, 100, 50, 0);
        break;
    case 1:
        verifyTouchEvent("touchmove", 1, 1, 1);
        verifyTouchPoint("touches", 0, 50, 50, 0);
        break;
    case 2:
        verifyTouchEvent("touchend", 0, 1, 0);
        verifyTouchPoint("changedTouches", 0, 50, 50, 0);
        break;

    default: testFailed("Wrong number of touch events! (" + which + ")");
    }
}

function touchTest1Element() {
    eventSender.addTouchPoint(100, 50);
    eventSender.touchStart();

    eventSender.updateTouchPoint(0, 50, 50);
    eventSender.touchMove();

    eventSender.releaseTouchPoint(0);
    eventSender.touchEnd();
}

if (window.layoutTestController)
    layoutTestController.waitUntilDone();

runTouchTest(function() {
    description("Test elements of height 0 receive touch events");
    lastEvent = null;
    touchTest1Element();
});

</script>
</body>
</html>
