kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/platform/android/fast/events/touch/touch-child-inside-empty-element-expected.txt

Test elements of height 0 receive touch events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS lastEvent.type is "touchstart"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.touches[0].target.id is "test1"
PASS lastEvent.touches[0].pageX is 100
PASS lastEvent.touches[0].pageY is 50
PASS lastEvent.touches[0].clientX is 100
PASS lastEvent.touches[0].clientY is 50
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.changedTouches[0].pageX is 100
PASS lastEvent.changedTouches[0].pageY is 50
PASS lastEvent.changedTouches[0].clientX is 100
PASS lastEvent.changedTouches[0].clientY is 50
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.targetTouches[0].pageX is 100
PASS lastEvent.targetTouches[0].pageY is 50
PASS lastEvent.targetTouches[0].clientX is 100
PASS lastEvent.targetTouches[0].clientY is 50
PASS lastEvent.targetTouches[0].identifier is 0
PASS lastEvent.type is "touchmove"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.touches[0].pageX is 50
PASS lastEvent.touches[0].pageY is 50
PASS lastEvent.touches[0].clientX is 50
PASS lastEvent.touches[0].clientY is 50
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.type is "touchend"
PASS lastEvent.touches.length is 0
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 0
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.changedTouches[0].pageX is 50
PASS lastEvent.changedTouches[0].pageY is 50
PASS lastEvent.changedTouches[0].clientX is 50
PASS lastEvent.changedTouches[0].clientY is 50
PASS lastEvent.changedTouches[0].identifier is 0
PASS successfullyParsed is true

TEST COMPLETE
